Handle CustomerBOImpl Exceptions

+++ back-end/src/main/java/lk/ijse/dep10/pos/business/custom/impl/CustomerBOImpl.java
@@ -1,6 +1,8 @@
 package lk.ijse.dep10.pos.business.custom.impl;
 
 import lk.ijse.dep10.pos.business.custom.CustomerBO;
+import lk.ijse.dep10.pos.business.exception.BusinessException;
+import lk.ijse.dep10.pos.business.exception.BusinessExceptionType;
 import lk.ijse.dep10.pos.business.util.Transformer;
 import lk.ijse.dep10.pos.dao.DAOFactory;
 import lk.ijse.dep10.pos.dao.DAOType;
@@ -11,7 +13,6 @@ import lk.ijse.dep10.pos.dto.CustomerDTO;
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class CustomerBOImpl implements CustomerBO {
@@ -30,6 +31,9 @@ public class CustomerBOImpl implements CustomerBO {
         try (Connection connection = dataSource.getConnection()) {
             customerDAO.setConnection(connection);
 
+            if (customerDAO.existsCustomerByContact(customerDTO.getContact())) {
+                throw new BusinessException(BusinessExceptionType.DUPLICATE_RECORD, "Save failed: Contact number: " + customerDTO.getContact() + " already exists");
+            }
 
             return transformer.fromCustomerEntity(customerDAO.save(transformer.toCustomerEntity(customerDTO)));
         }
@@ -40,6 +44,13 @@ public class CustomerBOImpl implements CustomerBO {
         try (Connection connection = dataSource.getConnection()) {
             customerDAO.setConnection(connection);
 
+            if (customerDAO.existsCustomerByContactAndNotId(customerDTO.getContact(), customerDTO.getId())) {
+                throw new BusinessException(BusinessExceptionType.DUPLICATE_RECORD, "Update failed: Contact number: " + customerDTO.getContact() + " already exists");
+            }
+
+            if (!customerDAO.existsById(customerDTO.getId()))
+                throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                        "Update failed: Customer ID: " + customerDTO.getId() + " does not exist");
 
             customerDAO.update(transformer.toCustomerEntity(customerDTO));
         }
@@ -51,17 +62,28 @@ public class CustomerBOImpl implements CustomerBO {
             customerDAO.setConnection(connection);
             orderCustomerDAO.setConnection(connection);
 
+            if (orderCustomerDAO.existsOrderByCustomerId(customerId)) {
+                throw new BusinessException(BusinessExceptionType.INTEGRITY_VIOLATION, "Delete failed: Customer ID: " + customerId + " is already associated with some orders");
+            }
+
+            if (!customerDAO.existsById(customerId))
+                throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                        "Delete failed: Customer ID: " + customerId + " does not exist");
+
+
             customerDAO.deleteById(customerId);
         }
     }
 
     @Override
-    public Optional<CustomerDTO> findCustomerByIdOrContact(String idOrContact) throws Exception {
+    public CustomerDTO findCustomerByIdOrContact(String idOrContact) throws Exception {
         try (Connection connection = dataSource.getConnection()) {
             customerDAO.setConnection(connection);
 
             return customerDAO.findCustomerByIdOrContact(idOrContact)
-                    .map(transformer::fromCustomerEntity);
+                    .map(transformer::fromCustomerEntity)
+                    .orElseThrow(() -> new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                            "No customer record is associated with the id or contact: " + idOrContact));
         }
     }
